checkpoint, moving some things around config name collision still breaking functionality for now
