+++ src/dynode/__init__.py
@@ -8,19 +8,19 @@ DynODE is currently under active development and will be substantially
 refactored in the near future!
 """
 
-from . import typing as typing
 from . import utils, vis_utils
-from ._simulation_date import SimulationDate
 
 # keep imports relative to avoid circular importing
 from .abstract_initializer import AbstractInitializer
 from .abstract_parameters import AbstractParameters
 from .config import Config
+from .config.simulation_date import SimulationDate
 from .covid_sero_initializer import CovidSeroInitializer
 from .dynode_runner import AbstractDynodeRunner
 from .mechanistic_inferer import MechanisticInferer
 from .mechanistic_runner import MechanisticRunner
 from .static_value_parameters import StaticValueParameters
+from .typing import typing as typing
 
 # Defines all the different modules able to be imported from src
 __all__ = [
