+++ src/dynode/config/__init__.py
@@ -1,23 +1,24 @@
 """DynODE configuration module."""
 
 from .bins import AgeBin, Bin, DiscretizedPositiveIntBin, WaneBin
-from .config_definition import (
-    Compartment,
-    Initializer,
-    SimulationConfig,
-)
+from .compartment import Compartment
 from .dimension import (
     Dimension,
     FullStratifiedImmuneHistoryDimension,
     LastStrainImmuneHistoryDimension,
     VaccinationDimension,
 )
-from .inference import InferenceProcess, MCMCProcess, SVIProcess
+from .initializer import Initializer
 from .params import (
     Params,
     SolverParams,
     TransmissionParams,
 )
+from .pre_packaged import SIRConfig, SIRInferedConfig
+from .simulation_config import (
+    SimulationConfig,
+)
+from .simulation_date import SimulationDate
 from .strains import Strain
 
 __all__ = [
@@ -36,7 +37,7 @@ __all__ = [
     "Params",
     "SolverParams",
     "TransmissionParams",
-    "InferenceProcess",
-    "MCMCProcess",
-    "SVIProcess",
+    "SIRConfig",
+    "SIRInferedConfig",
+    "SimulationDate",
 ]
