+++ src/dynode/config/params.py
@@ -17,7 +17,7 @@ from pydantic import (
 )
 from typing_extensions import Self
 
-from ..typing import DeterministicParameter
+from ..typing.typing import DeterministicParameter
 from .strains import Strain
