+++ src/dynode/config/pre_packaged/__init__.py
@@ -1,0 +1,5 @@
+"""A module containing pre-packaged Configs for easy off the shelf use."""
+
+from .example_sir_config import SIRConfig, SIRInferedConfig
+
+__all__ = ["SIRConfig", "SIRInferedConfig"]
