+++ src/dynode/infer/deterministic_parameter.py
@@ -1,70 +1,6 @@
-"""Module for declaring types to be used within DynODE config files."""
+"""A deterministic parameter thats value depends on a not yet realized value."""
 
-from typing import Any, Callable, Optional, Tuple
-
-import jax
-import numpyro.distributions as dist
-from annotated_types import Ge, Le
-from jaxtyping import PyTree
-from typing_extensions import Annotated, Union
-
-CompartmentState = Tuple[jax.Array, ...]
-CompartmentGradients = Tuple[jax.Array, ...]
-
-SEIC_Compartments = Tuple[
-    jax.Array,
-    jax.Array,
-    jax.Array,
-    jax.Array,
-]
-# a timeseries is a tuple of compartment sizes where the leading dimension is time
-# so SEIC_Timeseries has shape (tf, SEIC_Compartments.shape) for some number of timesteps tf
-SEIC_Timeseries = Tuple[
-    jax.Array,
-    jax.Array,
-    jax.Array,
-    jax.Array,
-]
-
-UnitIntervalFloat = Annotated[float, Ge(0.0), Le(1.0)]
-
-
-ODE_Eqns = Callable[
-    [jax.typing.ArrayLike, CompartmentState, PyTree],
-    CompartmentGradients,
-]
-
-ObservedData = Union[Tuple[jax.Array, ...], jax.Array]
-
-
-class SamplePlaceholderError(Exception):
-    """A special error raised if you attempt to randomly sample a placeholder variable."""
-
-    pass
-
-
-class PlaceholderSample(dist.Distribution):
-    """A parameter that draws its values from an external set of samples."""
-
-    def __init__(self):
-        """Create a PlaceholderSample distribution."""
-        super().__init__()
-
-    def sample(self, _, sample_shape=()):
-        """Retrieve sample from an external set of samples.
-
-        Raises
-        ------
-        SamplePosteriorError
-            if sample is called outside of an in-place substitute context like
-            numpyro.handlers.substitute() or numpyro.infer.Predictive.
-        """
-        raise SamplePlaceholderError(
-            "Attempted to sample a PosteriorSample parameter outside of a "
-            "Predictive() context. This likely means you did not provide "
-            "posterior samples to the context via numpyro.infer.Predictive() or "
-            "numpyro.handlers.substitute()."
-        )
+from typing import Any, Callable, Optional
 
 
 class DeterministicParameter:
