--- src/dynode/model_configuration/pre_packaged/__init__.py
@@ -1,6 +1,0 @@
-"""A module containing pre-packaged Configs for easy off the shelf use."""
-
-from .covid_seip_config import SEIPCovidConfig
-from .example_sir_config import SIRConfig
-
-__all__ = ["SEIPCovidConfig", "SIRConfig"]
