--- src/dynode/model_odes/__init__.py
@@ -1,5 +1,0 @@
-"""Module to hold all DynODE ODE flows."""
-
-from .seip_model import seip_ode
-
-_all_ = [seip_ode]
