+++ tests/test_integration.py
@@ -22,7 +22,7 @@ from dynode import (
     MechanisticRunner,
     StaticValueParameters,
 )
-from dynode.model_odes import seip_ode
+from dynode.simulate import seip_ode
 
 CONFIG_GLOBAL_PATH = "tests/test_config_global.json"
 INITIALIZER_CONFIG_PATH = "tests/test_config_initializer.json"
